Simplificados ficheros JavaScript
