--- src/components/checkHTTPS.js
@@ -1,4 +1,0 @@
-export function checkHTTPS( ) {
-    var protocol = window.location.protocol;
-    return protocol;
-}
