--- src/components/detectLanguaje.js
@@ -1,4 +1,0 @@
-export function detectLanguaje() {
-    var language = navigator.language;
-    return language;
-}
